+++ src/OCBridgeServiceProvider.php
@@ -42,6 +42,7 @@ class OCBridgeServiceProvider extends ServiceProvider
     {
         if ($this->app->runningInConsole()) {
             $this->commands([
+                Console\InstallCommand::class,
                 Console\MessageCommand::class,
             ]);